Initial Exactly-Once Execution (#3)
This PR contains a preliminary implementation of exactly-once execution.
Every workflow invocation is associated with an idempotency key, or
unique ID. This can be passed by the client as a parameter when the
workflow is invoked or can be generated automatically by Operon. If a
workflow is executed multiple times with the same idempotency key, all
successful executions return the same result and subsequent executions
after the first successful execution have no effect on the database, so
the effect is the same as if each function execution in the workflow
occurred exactly once.

To implement this, we record the output of each function execution in a
workflow to the database, serialized to JSON. If the workflow is later
re-executed, we return the recorded outputs for each
previously-completed function execution. Additionally, we record the
workflow input during the first execution and during re-executions use
the originally recorded input, to semantically clarify a situation where
a function is re-executed with the same idempotency key but different
inputs.

Next steps:

- Automatic recovery. If an Operon executor crashes, upon restart it
should automatically complete any incomplete workflows without needing
the client to resend them.
- Reduced overhead. We should be able to avoid adding writes to
read-only transactions, perhaps by deferring recording of their outputs
to the next transaction with a write. We may also be able to avoid the
additional transaction inserted at the start of every workflow to record
its input, perhaps by deferring the check to the first function that
performs a write.
- Garbage collection of recorded outputs.

---------

Co-authored-by: Qian Li <qian.li@dbos.dev>