Support External Calls (#5)
This PR implements a basic mechanism for external communication.

External API calls (or other non-deterministic operations) can be
wrapped in a "communicator" function and invoked from a workflow. We
guarantee that a communicator executes at least once and that once it
completes, it is not re-executed. We provide configurable automatic
retries for communicators.

As a next step, we may want to make our retry system more powerful, for
example by allowing users to define different retry policies for
different errors.

Also, could use suggestions for better names for the "Communicator"
class and "external" method.

---------

Co-authored-by: Peter Kraft <petereliaskraft@gmail.com>