Enhanced Error Handling (#9)
This PR improves error handling in Operon. We introduce a consistent
error handling model in Operon workflows: if an error is thrown in a
transaction or communicator, we check it against a list of retryable
errors, and if it is not retryable, we throw it to the caller.
Currently, in transactions we retry only serialization errors, while in
communicators we retry all errors. Before throwing an error to the
caller, we record it in the database (in the 'error' column of the
FunctionOutputs table) so re-executions of the workflow
deterministically throw the same errors.

In particular:

- Implemented a retry loop for transactions which currently only retries
serialization errors.
- Changed "idempotencyKey" to "workflowUUID" everywhere.
- Use operonNull as a special internal "null" return for checkExecution.
- For functions and communicators, we record either output or error, and
throw the error instead of returning null/undefined.
- Add more failure testing.

---------

Co-authored-by: Peter Kraft <petereliaskraft@gmail.com>