Update Function Recording (#13)
This PR makes changes to function recording and our enforcement of OAOO.
We guarantee that:

- If a workflow is run multiple times with the same UUID, subsequent
executions have no effect on the database or outside world beyond that
of the first. In other words, multiple executions with the same UUID are
safe and the workflow always runs once and only once.

Multiple workflow executions with the same UUID are not guaranteed to
return the same result (although they typically will).

In terms of performance:

- We add one write to each transaction (to record its output) and two
additional writes to the workflow (one to record its input, one to
record its output).
- We perform these writes lazily, so we do not add additional
transactions to a workflow, nor do we convert read-only transactions
into write transactions.

The main contribution of this PR is the lazy writes. Here's how they
work:

- When we start Operon, we create a global in-memory buffer of workflow
outputs.
- When we start a workflow, we create an in-memory buffer of function
outputs, then immediately add the workflow's input arguments to the
buffer.
- When we execute an operation that has no side effects (e.g., a
read-only transaction), we add its output to the workflow buffer.
- When we execute an operation that has side effects (e.g., a write
transaction or an external call), we add its output to the workflow
buffer then flush the workflow buffer to the database as part of the
operation's transaction.
- When a workflow completes, we add its output to the global buffer.
- Periodically, we flush the global buffer.

It is important to identify conflicts between simultaneous workflow
executions with the same UUID and distinguish them from other errors. To
do this, we perform the buffer flush at the beginning of any operation
with a side effect and include a placeholder value corresponding to the
operation. This placeholder acts as a guard preventing other executions
with the same UUID from making forward progress. If the operation
commits successfully, simultaneous executions fail with an OperonError.
In the future, we will fix this so those simultaneous executions instead
block until the "main" execution completes, then return its result.

Lazy recording improves performance massively for small workflows. We
quickly measured this on the Shop demo: end-to-end latency for an HTTP
GET request to the `api/products` endpoint dropped from 8ms to 1.3ms.
Previously, it required three transactions:

- Record workflow inputs.
- Look up all products, record the output.
- Record the workflow output. 

Now, it runs as just one read-only transaction and the workflow output
is recorded asynchronously.

---------

Co-authored-by: Qian Li <qian.li@dbos.dev>