Workflow Status Handling (#18)
This PR implements workflow status management. Invoking a workflow now
synchronously returns a `WorkflowHandle`. A handle to an existing
workflow (active or completed) can also be retrieved using the
workflow's UUID. Handles provide methods for checking a workflow's
status(`PENDING`, `SUCCESS`, or `ERROR`) and for retrieving a workflow's
output (blocking until it completes). A workflow's status is set to
`PENDING` sometime after it starts, but always before it performs an
operation with a side effect, such as a database write. Its status is
updated to `SUCCESS` or `ERROR` after it completes; at this point the
output is retrievable.

Our current implementation sets workflow status to `PENDING` as part of
flushing the function results buffer, which occurs every time we perform
an operation with a side effect. Updating the status to `SUCCESS` or
`ERROR` is done asynchronously alongside the workflow outputs buffer. A
special case in our current implementation is that top-level
transaction, send, and recv operations with side effects would be marked
as `SUCCESS` or `ERROR` as part of flushing the function results buffer.

To make behavior more intuitive, a workflow that is found to be
conflicting with another execution with the same UUID will no longer
fail, but will rather retrieve a handle on the successful execution and
await and return its result.

We additionally implement a very basic workflow recovery system. When an
Operon instance starts, it checks the database for workflows stuck in a
`PENDING` state and restarts them. Thus, if an executor crashes, its
replacement will pick up where it left off. In the future (particularly
in a distributed setting), we can make this more sophisticated.

---------

Co-authored-by: Peter Kraft <petereliaskraft@gmail.com>