Refactor System Database (#22)
This PR encapsulates all database interactions which do not need to be
in the same database as user data into a new module called
`SystemDatabase`, which has an abstract interface and a Postgres-based
implementation. The goal is to make it easier for us to adopt a
different system database backend (e.g., FoundationDB) or a different
user database interface (e.g., TypeORM/Sequelize).

We have encapsulated all Operon database interactions into
`SystemDatabase` (including workflow status management and the
notifications system) except for the `TransactionOutputs` table, which
must be managed as part of user database transactions to guarantee OAOO
semantics.

Because workflow status can no longer be managed as part of user
transactions, we have altered its semantics. The `PENDING` workflow
status has been eliminated. Instead, the status of a workflow is
`UNKNOWN` until it completes, at which point it is set to `SUCCESS` or
`ERROR`. We may re-introduce the `PENDING` state in the future if there
is a need for it and we find a way to implement it efficiently.

We have also temporarily removed workflow metadata garbage collection.
This garbage collection is tricky because once a workflow's metadata is
deleted, we can no longer guarantee OAOO for the workflow (i.e.,
re-executions with the same UUID are no longer safe). Our past
implementation leveraged complicated transactions to garbage-collect
some metadata for a workflow but not all of it. In the future, we will
introduce a more principled interface and implementation that
garbage-collects all workflow metadata and makes it clear that workflow
re-execution is not safe after garbage collection.

---------

Co-authored-by: Peter Kraft <petereliaskraft@gmail.com>