Assume User Database Exists (#26)
This PR modifies Operon initialization to assume the user database
already exists instead of trying to create it. The user database should
always exist because Operon is designed for use with a pre-existing
database and schema; it does not create the user's tables. This change
simplifies initialization and eliminates the need to connect to the
default `postgres` database, which may require elevated privilege.

---------

Co-authored-by: Peter Kraft <petereliaskraft@gmail.com>