FoundationDB Support (#32)
This PR implements FoundationDB as an alternative system database for
Operon.

We don't want to introduce a hard dependency on the FoundationDB JS
client,
[`node-foundationdb`](https://github.com/josephg/node-foundationdb),
because it requires the FoundationDB client header files. Therefore, we
make `node-foundationdb` an optional dependency in `package.json` so it
is not automatically installed on computers that don't have the
FoundationDB client installed.

Additionally, we exclude all FoundationDB code from `npm run build`,
`npm run lint`, and `npm run test` by default. To include it, run `npm
run build-fdb`, `npm run lint-fdb`, and `npm run test-fdb`.

---------

Co-authored-by: Qian Li <qian.li@dbos.dev>