Export Postgres transaction ID (#41)
This is a small PR that obtains the transaction ID of non read-only
transactions and adds it as an attribute to the tracing span. This
information will be useful to our provenance system where we can link
each transaction back to its workflowUUID and function ID.

To see the effect, you need to enable `JAEGER_EXPORTER` in Operon
config, start a Jaeger server, and run some tests. Then, you'll be able
to see `postgres_transaction_id` for transactions that are not marked as
read-only.