Basic provenance (#42)
This PR implements a basic provenance daemon in a standalone module. The
provenance daemon leverages Postgres logical replication to capture
changes to all user tables in the user database, exporting them to the
observability database.

The provenance daemon is simply a Typescript object, it does not need to
be in the same process as Operon:
```javascript
daemon = new ProvenanceDaemon(operonConfig, <logical replication slot name>);
await daemon.start();
```
The daemon will create a logical replication slot in the user database
and monitor changes through the slot.

The daemon has three requirements from the database:

1. The Postgres database must be configured with `wal_level = logical`.
This requires a restart. At least one open replication slot must be
available.
2. The `wal2json` plugin must be installed. It is installed by default
on most cloud databases, including RDS and Azure.
3. Updates and deletes can only be exported if their table's rows can be
uniquely identified, either because they have a primary key or a replica
identity.

Provenance can potentially capture sensitive user data; in the future we
must add a way to mask this data.

---------

Co-authored-by: Peter Kraft <petereliaskraft@gmail.com>