Refactor Endpoint Decorators (#55)
This PR aims to move HTTP server related decorators to a separate
module, so the decorators in the root directory are only for
operon-core.

Specifically:
- Refactor HTTP endpoint related decorators to a separate `handler.ts`
file in the httpServer module, including `ArgSource`, `ArgSources`,
`APITypes`, `GetApi`, and `PostApi`. These decorators are not directly
relevant to core Operon functionalities and only make sense when using
with HTTP handlers.
- Factor out `rawContext` from OperonContext to `koaContext` in
HandlerContext in the httpServer module, because it is Koa dependent and
should not belong to the base context class.
- Remove `origFunction` from `OperonMethodRegistrationBase` and use the
replacement function everywhere. The original function should only be
invoked from within the wrapper.

After this PR, we plan to move all tests to use decorators.