Operon HTTP Endpoint Refactoring (#60)
This PR aims to clean up the code around Operon HTTP server and handler
context.

This PR allows functions to set customized HTTP status code by throwing
`OperonResponseError`, and also adds a `log` function in
`HandlerContext` by leveraging the Operon telemetry system.

The current semantics for returning HTTP response:
- If functions return successfully and hasn't set the body, we set the
body to the function return value. The status code will be automatically
set to 200 or 204 (if the body is null/undefined).
- In case of an exception, we include the original error in the response
body and set status code accordingly:
- If an Operon throws a client-side error (e.g., data validation error),
we return 400.
- If an error contains a `status` field, we return the specified status
code.
  - Otherwise, we return 500.

This PR also refactored the OperonContext to only include a Node.js HTTP
request (IncomingMessage) and removed the response field.

We still keep the raw Koa context in HandlerContext because we do not
want to reinvent another HTTP server. The raw Koa context can only be
accessed inside non-WF/TX/Comm handlers because the response object is
not logged for OAOO.