Remove global method registration (#61)
replaces methodRegistry in decorators.ts with a per Operon instance
method registry. Classes are passed to operon.init in order to be
registered.

Collection of OperonMethodRegistrations is passed to PostgresExporter
via init parameter. Other exports ignore this parameter.

TelemetryCollector.init now takes collection of method registrations in
order to pass to the exporters

Http server iterates over the collection of registered operations in the
provided operon instance in registerDecoratedEndpoints

~Not 100% sure on tests. Will see what the CI server does with them~ I
had an issue where I was failing to destroy the operon object. Fixed