Middleware Module (#67)
This PR refactors HTTP middleware related interfaces and moves them a
separate `middleware.ts` file. The goal is to explicitly define the
input and output types of the HTTP Auth middleware. An HTTP authenticate
middleware takes in a `MiddlewareContext` which contains the raw Koa
context and required roles to run a target endpoint, and returns an
object which contains the authenticated user and roles.