Operon Deploy (#68)
Implements a single command to deploy a serverless Operon application to
a bare-metal machine:

```
npx operon deploy --name <app name>
```

For now, this command must be run on a bare-metal machine with the cloud
API active. Docs and support infrastructure are here:
https://github.com/dbos-inc/operon-cloud/pull/33

I'm worried it may be confusing to have the same CLI manage both local
development commands like `npx operon start` and `npx operon init` and
cloud deployment commands like `npx operon deploy` and eventually more
complex commands to log in and manage deployments. In the future, we may
want to separate the local Operon CLI, which should remain in this
repository, from the cloud CLI, which may belong in a dedicated
open-sourcable repository.