offer a nicer abstraction for logging (#73)
Refine the log abstraction so users can call `context.info` and such,
directly. (Mostly a mapping of what the native `console`
[offers](https://developer.mozilla.org/en-US/docs/Web/API/console).)

Also add the name of the user function, and the location in the
transpiled code, of the log call.

Example with the hello example:

```
  @OperonTransaction()
  static async helloFunction (txnCtxt: TransactionContext, name: string)  {
    const greeting = `Hello, ${name}!`
    const { rows } = await txnCtxt.pgClient.query<{greeting_id: number}>("INSERT INTO OperonHello(greeting) VALUES ($1) RETURNING greeting_id", [greeting])
    txnCtxt.warn(`Inserted greeting ${rows[0].greeting_id}: ${greeting}`)
    return `Greeting ${rows[0].greeting_id}: ${greeting}`;
  }

  @OperonWorkflow()
  @GetApi('/greeting/:name')
  static async helloWorkflow(workflowCtxt: WorkflowContext, name: string) {
    workflowCtxt.info("Hello, world!");
    return await workflowCtxt.transaction(Hello.helloFunction, name);
  }
```


![Screenshot 2023-09-26 at 17 41
12](https://github.com/dbos-inc/operon/assets/3437048/b8137600-d24c-4355-9b5a-5b6e25e94696)


EDIT:

This PR became more involved due to lifting the log methods in the base
`OperonContext`. We were exporting telemetry definitions from
`src/telemetry/index.ts`, which created a circular dependency when
adding an `import { Logger } from "telemetry/"` statement in
`src/context.ts`. This is now fixed by the full removal of
`src/telemetry/index.ts` and appropriate update of import statements
throughout the code.