update Operon configuration processing (+ interpolate environment variables) (#79)
Today, the `Operon` object is responsible, in the absence of an input
`OperonConfig` for setting its own configuration. This is done by
looking up for a configuration file at a known location.

The idea of allowing an optional `OperonConfig` argument was to enable
programmatic adjustment of the configuration, most specifically in the
case of unit testing. In most use cases, `Operon` users would simply
call `Operon()` and let the SDK figure out its own configuration using
the configuration file.

Today, two things have changed:
- Users do not call `Operon` directly
- Configuration is becoming more complex. The original use case for this
PR was to support interpolating environment variables to help users.

This PR takes a firm approach toward our new serverless model and:
- Make `OperonConfig` mandatory in the `OperonConstructor`
- Move the responsibility of parsing and creating this configuration to
the `operon-runtime`. The runtime is the only execution path expected to
build an `Operon` object, so it makes sense to let it prepare the
configuration
- Use a standard tool, [envsubst](https://linux.die.net/man/1/envsubst),
to interpret environment variables
- Added a helper to `OperonContext` to retrieve application
configuration variables
- Add types for the runtime configuration and separate them from the
`Operon` core configuration