Workflow Recovery (#90)
This PR implements automatic workflow recovery. Whenever Operon starts,
it finds all workflow executions that have started but not completed and
runs them to completion. Specifically, we:

1. Synchronously set a workflow's status to `PENDING` at workflow start
time, so if a workflow has started its status is guaranteed to be
durably set in the system database.
2. Add new fields to workflow status, including `authenticatedUser`. 
3. At Operon startup time, automatically run all workflows in a
`PENDING` state to completion by restarting them from the beginning.

Next steps & discussion:
1. We shouldn't be passing around the `IncomingMessage` object because
it's stateful in a nasty way (it includes a socket). Will figure out a
solution in the next PR.
2. Since `operon.workflow` becomes async, we now have to write: `await
(await ctxt.invoke(Class).wf(...)).getResult()`

---------

Co-authored-by: Qian Li <qian.li@dbos.dev>
Co-authored-by: Chuck B <chuck.bear@dbos.dev>