Fix Handler Context Bug (#93)
This PR fixes a bug in handler context: it didn't pass the handler
context as a parent context to invoked transactions. Added a test to
make sure that child workflows and transactions can get the correct auth
information from parentCtx.

It also fixes an issue in a previous PR merge that disabled the console
exporter. We should always have a default console exporter unless
specify something else.