Workflow invoking child workflows (#97)
This PR adds a `childWorkflow` method in workflow context that enables a
workflow to invoke child workflows and guarantees exactly-once for such
invocations. Now inside a workflow, you can write:
```javascript
handle = await wfCtxt.childWorkflow(ClassName.workflowName, ...args...);
```

We pass in the parent context and assign each child workflow a
deterministic workflow UUID which appends the current function ID to the
parent's workflow UUID. Therefore, when we rerun a workflow, child
workflows can also be deterministically retried.

The initial plan was to support childWorkflow through `ctxt.invoke()`,
however, it can potentially cause a recursive type inference issue that
is very hard to debug: if a workflow's return type depends on
invoke(ThisClass), the invoke(ThisClass) return type in turn depends on
the return type of the workflow, so`tsc` may not be able to infer the
return type due to recursion (unless we explicitly specify the return
type for every workflow in the class).