Operon Logger (#101)
This PR adds a proper logger to Operon. The property we desire are:
- Structured logging
- Configurable log levels
- Operon framework logs + user code logs
- Not creating a new logger per operation/request
- Optionally add context information in log entries

This PR *doesn't* consider exporting to observability backends (e.g.,
ours, Splunk). I think the work of exporting log entries fits better in
another PR.

To meet our desire, this PR:
- Replaces the `Logger` in `src/telemetry/logs.ts` with a wrapper around
[winston](https://www.npmjs.com/package/winston).
- Add a `logger` property to `OperonConfig` and `OperonRuntimeConfig`.
This logger is a Winston logger.
- Add a `logger` property to `OperonContext`. This logger is our wrapper
around Winston.
- Instantiating an `OperonContext` creates a new wrapper instance, which
itself holds a reference to the global logger
- The wrapper expose Winston's methods (`info()`, `warn()`, `error()`,
etc). Users can request the addition of contextual information to log
entries by passing an optional `true` flag, after their log message.
- Our formatter's `log.error()` expects an error object, not a string.
This is so we can capture the stack trace properly. We can discuss this.

This new logger subsystem replaces the console exporter, which job was
simply to call `console.log`. Finally, this PR replaces calls to the
`console` across the code with calls to the logger.

The expected usage is as follows:

```ts
@OperonTransaction()
static async helloTransaction(txnCtxt: KnexTransactionContext, name: string) {
  const logger = txnCtxt.getLogger();
  const greeting = `Hello, ${name}!`
  logger.info('log message with context data', true);
  logger.warn('warning without context data');
  logger.error(new Error('oops'));
  logger.error(new Error('oops with context data'), true);
  ...
}
```

Example output:

![Screenshot 2023-10-06 at 16 43
25](https://github.com/dbos-inc/operon/assets/3437048/85250ea1-eae7-4e01-90e1-d8cf2621388f)



Others:
- Modify Operon and Runtime configurations to definitively happen in
`src/operon-runtime/config.ts`