JSON Schema for operon-config (#142)
The [RedHat YAML VSCode
extension](https://marketplace.visualstudio.com/items?itemName=redhat.vscode-yaml)
supports YAML validation via [JSON Schema](https://json-schema.org/).

I used
[ts-json-schema-generator](https://github.com/vega/ts-json-schema-generator)
to generate the initial schema then simplified it by hand.

> Note, I couldn't get ts-json-schema-generator to work directly against
the operon repo, so I copied the `ConfigFile` and types it depends on
out into a separate scratch project to generate the initial schema.

Assuming you have the YAML extension installed, you can specify the
schema in your operon-config file by adding this comment at the top

```yaml
# yaml-language-server: $schema=<URL to schema>
```

FYI, I'm opening this PR as a draft because a) I'm not sure we want to
do this now and b) we would need to pick a URL to publish the schema so
that we can included it in the hello example. We could use a
`raw.githubusercontent.com` URL or publish it somewhere on the main dbos
website.