Operon CLI login (#150)
This PR implements `npx operon login -u <username>`, which currently
takes in a user name and creates a "valid but fake" JWT token to be used
in `npx operon deploy`. This PR removes the previously hard coded JWT
token in deploy.

`login` stores the generated token in a local file
`.operon/credentials`, and `deploy` reads the token from that file.
In the future, we will integrate with a 3rd-party auth system (e.g.,
Okta) and we will re-direct users to login through the auth system and
obtain a real JWT token there.