Create an npm publish action (#155)
This Github action publishes a package to npm with a proper tag:
- Only `release/*` branches will be tagged as `latest`.
- Main branch will be tagged as `preview`.
- Publishes from other branches will be tagged as `test`.

When we run `npm install @dbos-inc/operon` it will only install the
latest release version, not the `preview` or `test` versions. Please the
published versions here:
https://www.npmjs.com/package/@dbos-inc/operon?activeTab=versions

This action currently can only be triggered manually.

Another cool thing about `JS-DevTools/npm-publish@v3` action is that it
can automatically skip a publish if the version already exists on NPM.
So we don't need to worry about version conflicts.