Database access to middleware (#154)
Provide additional functions for middleware, including database access,
logging, and configuration information.

This revealed a need for a better client interface to queries, that was
more in line with the experience expected for those clients. This is
added in a similar manner to transaction functions.

Tests for Knex, Prisma, TypeORM.  PG node is not intended to be used at this time.