OpenAPI support (#147)
Start of the work to add OpenAPI/Swagger support. 

So far, this PR only contains the code for reading the operation
typescript source and looking for classes, method, parameters and
decorators. I've added a `generate` verb to the Operon CLI that parses
the TS for the provided entrypoint file (Note, this is the TS entry
point not the JS entrypoint declared in the config file)

We will want to consider breaking the Operon runtime and CLI into
separate packages after we add this PR. Parsing the TS code means taking
a dependency on the `typescript` package. That's fine for the CLI tool
running on the user's machine (they need that package for compiling
their code anyway) but I'm guessing we don't want to require the
`typescript` dependency for runtime execution.