Selective Workflow Recovery (#173)
Previously, Operon tries to recover all workflows with `PENDING` status,
which is not scalable to the cloud. This PR implements selective
workflow recovery, and fixes a concurrency bug where the recovery may
interfere with normal executions.

This PR also adds a reserved HTTP POST endpoint
(`/operon-workflow-recovery`), allowing us to trigger workflow recovery
externally. This endpoint will be transparent to users when we deploy
apps to the cloud.

Implementation details:
- We add an `executor_id` field in the `workflow_status` table that
records the ID of the machine that executes this workflow. The
`executor_id` field by default is "local" if Operon is tested locally,
and is set to the Firecracker VM ID if running on the cloud.
- During Operon init, we still recover workflows marked as "local", so
we can guarantee workflows always run to completion after local
restarts.
- We extract the executor ID from an `operon-executorid` HTTP header
(will be automatically set by the cloud), and record it when we init the
workflow status record.
- We then update the `recoverPendingWorkflows` method to accept a list
of executor IDs and only recover workflows that belong to those
executors.
- We update recovery tests for both PG and FDB system databases.