add cloud CLI applications subcommand to update an application (#179)
A subcommand to update applications.

For now, does the most simple thing: ask users to always provide the
full intent of the app record, i.e., provide a name and max_vms. For
now, only `max_vms` can be updated. Updating the application name
require more uplifting in the backend, which we should not do now.