Initial Debugger (#197)
This PR adds initial support for the debug mode. The main goal is to
faithfully replay past workflow executions. Under debug mode, everything
should be read-only: we skip initialization steps in the database and
don't record for OAOO. Instead of connecting to the actual user DB, we
connect to a debugger proxy which transforms queries to perform time
traveling.

Details:
- Add a CLI option for `npx dbos-sdk start`: `-d, --debug` with the URL
of the debug proxy.
- We can only execute a workflow in debug if we can find recorded
workflow status and transaction output. Otherwise, we throw an error.
- Create a new workflow context called `WorkflowContextDebug`, which
implements debugger specific methods. We currently only support
executing transaction functions or child workflows. Will implement more
in the future.
- Add a debugger test. We currently connect to the actual user DB in
debug mode, but once the proxy is ready, we can test with it.