Improve Transaction Execution (#205)
This PR makes two improvements to transaction execution: 

1. Fix an issue where transaction snapshot information was sometimes not
recorded for read-only transactions because the transaction results
buffer is only flushed by write transactions. Thus, snapshot information
for a read-only transaction was only recorded if its workflow contains a
subsequent write. To fix this, we now flush the transaction results
buffer asynchronously at the end of a workflow.

2. Only check if a transaction previously executed if the workflow UUID
is preset (Qian's idea!). Otherwise, the UUID is newly generated within
the SDK, so simultaneous executions with the same UUID should be
impossible. This reduces the overhead of the SDK to one additional
round-trip per transaction.