Fix Debug Mode Output Check (#223)
This PR solves a minor issue for the debug mode: If a function returns
nothing, we record `null` in the database (there's no good way to
serialize undefined to a valid JSON string). So when we get the recorded
output, we would get `null` instead of `undefined`. Therefore, the
debugger would incorrectly print out an error message when it compares
`undefined` to `null` when the function returns nothing.

This PR updates the debug mode code to handle the `undefined` return
value.