Send Telemetry Data Faster (#228)
Per our slack exchange with @maxdml , reduce the sleep wait time to
100ms and, every cycle, send all the data that has accumulated in the
queue. Thus there is no "max message size" and we send everything we
have to avoid delays and a growing RAM footprint.

I tested this change by re-running test_load and verifying that there
are exactly 10,000 unique traces in the monitoring DB as soon as the
test finishes. I'm considering modifying that test to check that.
Requires a companion PR on the dbos-cloud side that will ensure that
Zazu and otel-pg-exporter only use one postgres connection.

Ran all the other integration tests too.

Co-authored-by: Ubuntu <ubuntu@ip-10-0-0-75.ec2.internal>