Put ids into logs (#242)
The ts-side of the effort to gather app and vm IDs for monitoring.
Gather APPID and VMID from environment variables, if defined. Add them
to the attributes of both logs and traces (as applicationID and
executorID). Note that traces already uses executorID that it gets from
headers (for most spans). We only use the environment variable for those
spans where it is not already set.

---------

Co-authored-by: Ubuntu <ubuntu@ip-10-0-0-7.ec2.internal>
Co-authored-by: Ubuntu <ubuntu@ip-10-0-0-231.ec2.internal>