Fix System DB Creation in CLI (#244)
This PR fixes an issue where multiple applications may try to create the
same system DB and conflict with each other. This scenario is not common
in prod, but is common in our test_load.
This PR should prevent most of the conflicts, however, we cannot protect
concurrent `CREATE DATABASE`. So we must be careful about when we create
the system DB.