Ensure that logs are exported during error conditions (#259)
This fixes a few problems of the flavor "executor stopped but the log
entries are blank or missing":
1. The OTLP Log Exporter did not properly convert an Error object into a
log body. Changed to pass a string.
2. The DBOSRuntime.destroy() called some code to drain OTLP logs from
queues... but nobody called DBOSRuntime.destroy(). Assembled runtime
init/start_server under a single try/catch block and now call destroy on
error.
3. In the same destroy() path, we also needed to call `forceFlush()` on
the two OTLPExporters. Otherwise, rarely, things didn't get sent.
4. Added a signal handler for SIGTERM and SIGQUIT to log the fact that
we're exiting due to a signal. For a while, I thought about whether or
not this should be an error. In the end, decided that it's an `info`
because users may be sending SIGTERM during regular operations.